+++ src/lib/DatabaseClasses.cs
@@ -1,19 +1,14 @@
+using System;
+using System.Data;
+
 namespace HotelRezervacije
 {
-
-    //    return @"
-    //             CREATE TABLE rooms (id INTEGER PRIMARY KEY AUTOINCREMENT, name TEXT, capacity INTEGER NOT NULL, price_per_night DECIMAL(10, 2));
-    //             CREATE TABLE users (id INTEGER PRIMARY KEY AUTOINCREMENT, full_name TEXT NOT NULL, email TEXT NOT NULL UNIQUE, phone TEXT);
-    //             CREATE TABLE reservations (id INTEGER PRIMARY KEY AUTOINCREMENT, room_id INTEGER NOT NULL, user_id INTEGER NOT NULL, check_in DATE NOT NULL, check_out DATE NOT NULL, adults INTEGER NOT NULL, children INTEGER NOT NULL, FOREIGN KEY (room_id) REFERENCES rooms(id), FOREIGN KEY (user_id) REFERENCES users(id));
-    //             CREATE TABLE images (id INTEGER PRIMARY KEY AUTOINCREMENT, room_id INTEGER, image_data BLOB, FOREIGN KEY (room_id) REFERENCES rooms(id));";
-
-
     public class Room
     {
         public int Id { get; set; }
         public string Name { get; set; }
         public int Capacity { get; set; }
-        public float PricePerNight { get; set; }
+        public decimal PricePerNight { get; set; }
         public string Description { get; set; }
     }
 
@@ -39,6 +34,7 @@ namespace HotelRezervacije
         public int ReservationId { get; set; }
         public int GuestId { get; set; }
     }
+
     public class Reservation
     {
         public int Id { get; set; }
@@ -47,8 +43,9 @@ namespace HotelRezervacije
         public DateTime CheckIn { get; set; }
         public DateTime CheckOut { get; set; }
         public int NumberOfGuests { get; set; }
-        public float TotalPrice { get; set; }
+        public decimal TotalPrice { get; set; }
     }
+
     public class DatabaseImage
     {
         public int Id { get; set; }
@@ -62,5 +59,4 @@ namespace HotelRezervacije
         public string Name { get; set; }
         public string Icon { get; set; }
     }
-
 }
